testing changes

+++ common/common.go
@@ -81,7 +81,7 @@ func GetDir(c *cli.Context) string {
 }
 
 // BuildFunc bumps version and builds function.
-func BuildFunc(verbose bool, fpath string, funcfile *FuncFile, buildArg []string, noCache bool) (*FuncFile, error) {
+func BuildFunc(verbose bool, fpath string, funcfile *FuncFile, buildArg []string, noCache bool, retainDockerfile bool) (*FuncFile, error) {
 	var err error
 	if funcfile.Version == "" {
 		funcfile, err = BumpIt(fpath, Patch)
@@ -94,7 +94,7 @@ func BuildFunc(verbose bool, fpath string, funcfile *FuncFile, buildArg []string
 		return nil, err
 	}
 
-	if err := dockerBuild(verbose, fpath, funcfile, buildArg, noCache); err != nil {
+	if err := dockerBuild(verbose, fpath, funcfile, buildArg, noCache, retainDockerfile); err != nil {
 		return nil, err
 	}
 
@@ -102,7 +102,7 @@ func BuildFunc(verbose bool, fpath string, funcfile *FuncFile, buildArg []string
 }
 
 // BuildFunc bumps version and builds function.
-func BuildFuncV20180708(verbose bool, fpath string, funcfile *FuncFileV20180708, buildArg []string, noCache bool) (*FuncFileV20180708, error) {
+func BuildFuncV20180708(verbose bool, fpath string, funcfile *FuncFileV20180708, buildArg []string, noCache bool, retainDockerfile bool) (*FuncFileV20180708, error) {
 	var err error
 
 	if funcfile.Version == "" {
@@ -116,7 +116,7 @@ func BuildFuncV20180708(verbose bool, fpath string, funcfile *FuncFileV20180708,
 		return nil, err
 	}
 
-	if err := dockerBuildV20180708(verbose, fpath, funcfile, buildArg, noCache); err != nil {
+	if err := dockerBuildV20180708(verbose, fpath, funcfile, buildArg, noCache, retainDockerfile); err != nil {
 		return nil, err
 	}
 
@@ -150,7 +150,7 @@ func PrintContextualInfo() {
 	fmt.Println("Current Context: ", currentContext)
 }
 
-func dockerBuild(verbose bool, fpath string, ff *FuncFile, buildArgs []string, noCache bool) error {
+func dockerBuild(verbose bool, fpath string, ff *FuncFile, buildArgs []string, noCache bool, retainDockerfile bool) error {
 	err := dockerVersionCheck()
 	if err != nil {
 		return err
@@ -172,7 +172,11 @@ func dockerBuild(verbose bool, fpath string, ff *FuncFile, buildArgs []string, n
 		if err != nil {
 			return err
 		}
-		defer os.Remove(dockerfile)
+
+		if !retainDockerfile {
+			defer os.Remove(dockerfile)
+		}
+
 		if helper.HasPreBuild() {
 			err := helper.PreBuild()
 			if err != nil {
@@ -180,6 +184,7 @@ func dockerBuild(verbose bool, fpath string, ff *FuncFile, buildArgs []string, n
 			}
 		}
 	}
+
 	err = RunBuild(verbose, dir, ff.ImageName(), dockerfile, buildArgs, noCache)
 	if err != nil {
 		return err
@@ -194,7 +199,7 @@ func dockerBuild(verbose bool, fpath string, ff *FuncFile, buildArgs []string, n
 	return nil
 }
 
-func dockerBuildV20180708(verbose bool, fpath string, ff *FuncFileV20180708, buildArgs []string, noCache bool) error {
+func dockerBuildV20180708(verbose bool, fpath string, ff *FuncFileV20180708, buildArgs []string, noCache bool, retainDockerfile bool) error {
 	err := dockerVersionCheck()
 	if err != nil {
 		return err
@@ -216,7 +221,11 @@ func dockerBuildV20180708(verbose bool, fpath string, ff *FuncFileV20180708, bui
 		if err != nil {
 			return err
 		}
-		defer os.Remove(dockerfile)
+
+		if retainDockerfile {
+			defer os.Remove(dockerfile)
+		}
+
 		if helper.HasPreBuild() {
 			err := helper.PreBuild()
 			if err != nil {
@@ -224,6 +233,7 @@ func dockerBuildV20180708(verbose bool, fpath string, ff *FuncFileV20180708, bui
 			}
 		}
 	}
+
 	err = RunBuild(verbose, dir, ff.ImageNameV20180708(), dockerfile, buildArgs, noCache)
 	if err != nil {
 		return err
@@ -349,6 +359,9 @@ func Exists(name string) bool {
 }
 
 func writeTmpDockerfile(helper langs.LangHelper, dir string, ff *FuncFile) (string, error) {
+
+	fmt.Println("~~in writeTmpDockerfile dockerfile")
+
 	if ff.Entrypoint == "" && ff.Cmd == "" {
 		return "", errors.New("entrypoint and cmd are missing, you must provide one or the other")
 	}
@@ -501,12 +514,14 @@ func DockerPush(ff *FuncFile) error {
 		return err
 	}
 	fmt.Printf("Pushing %v to docker registry...", ff.ImageName())
+	/*
 	cmd := exec.Command("docker", "push", ff.ImageName())
 	cmd.Stderr = os.Stderr
 	cmd.Stdout = os.Stdout
 	if err := cmd.Run(); err != nil {
 		return fmt.Errorf("error running docker push, are you logged into docker?: %v", err)
 	}
+	 */
 	return nil
 }
 
@@ -516,6 +531,7 @@ func DockerPushV20180708(ff *FuncFileV20180708) error {
 	if err != nil {
 		return err
 	}
+	/*
 	fmt.Printf("Pushing %v to docker registry...", ff.ImageNameV20180708())
 	cmd := exec.Command("docker", "push", ff.ImageNameV20180708())
 	cmd.Stderr = os.Stderr
@@ -523,6 +539,8 @@ func DockerPushV20180708(ff *FuncFileV20180708) error {
 	if err := cmd.Run(); err != nil {
 		return fmt.Errorf("error running docker push, are you logged into docker?: %v", err)
 	}
+	 */
+	fmt.Printf("Pushing %v to docker registry...", ff.ImageNameV20180708())
 	return nil
 }
